backend/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
dj-database-url
gunicorn
drf-spectacular
django-cors-headers